CollisionSystem

+++ GameAPI/Player.cs
@@ -2,7 +2,7 @@
 {
     public class Player : GameObject
     {
-        public Player(GridLoader loader, long x, long y) : base(loader, x, y, Types.Player, Grids.Player)
+        public Player(GridLoader loader, int x, int y) : base(loader, x, y, Types.Player, Grids.Player)
         {
 
         }
